Safari-611.1.10/LayoutTests/http/tests/security/mixedContent/insecure-script-with-secure-cookies.html

<!DOCTYPE html>
<html>
<body>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<p>This test loads a secure iframe that reads secure cookies and then tries to load an insecure script.  This should block the insecure script since secure cookies were read.</p>
<iframe src="https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-script-secure-cookies.html"></iframe>
</body>
</html>
