LayoutTests/http/tests/security/mixedContent/insecure-script-with-secure-cookies-expected.txt

CONSOLE MESSAGE: line 3: secureCookie=yes
CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-script-secure-cookies.html was not allowed to run insecure content from http://127.0.0.1:8080/security/mixedContent/resources/script.js.

This test loads a secure iframe that reads secure cookies and then tries to load an insecure script. This should block the insecure script since secure cookies were read.
